requirements.txt

Flask
requests
Pillow
tqdm
flask-compress
